#19 Customers could not go back and edit their address details in the checkout process.
This issue seems to have been caused by a bug in Spree's source. When an update was performed the 'id' for the shipping and billing addresses was removed because 'id' was not considered a permitted field. Without the 'id' field a new address would be created rather than updating the original address. This interacted with our changes that moved first name, last name and phone out of the address forms and into the checkout process. The new addresses did not contain the personal information and that would cause validation to fail.