redirect_to override now always returns something.
The redirect_to override in checkout_controller_decorator.rb broke existing code in the CheckoutController that used the pattern "redirect_to 'somewhere' and return". The conditional logic after the redirect_to was not evaluated because redirect_to was returning a falsy value, nil in this case, and the 'and' keyword only evaluates if the left-hand-side code returns a truthy value. The original redirect_to returned the body content.

Fixes #27.